用户服务接口部分完成

+++ src/main/java/com/help/each/config/AppConfig.java
@@ -18,5 +18,5 @@ public class AppConfig {
     /**
      * 默认分页大小
      */
-    private Integer pageSize = 10;
+    private Long pageSize = 10L;
 }
